openenv_panacea/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "panacea-oversight-env"
version = "1.0.0"
description = "Adversarial hospital oversight OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "fastapi",
    "uvicorn",
]

[project.optional-dependencies]
training = [
    "unsloth",
    "trl",
    "matplotlib",
]
